S4 Ex 6-10

+++ Exercicios/Seção 4/ex10.py
@@ -1,2 +1,2 @@
-kmh = int(input("Insira uma velocidade em km/h: "))
+kmh = float(input("Insira uma velocidade em km/h: "))
 print(f"{kmh}km/h é igual a {kmh / 3.6}m/s")
